refactor(inspection-edit): extract 8 editor modals/drawers to app/components/editor/ (3307→~2300)
Behavior-preserving: inline modal JSX moved byte-identical into presentational
components (AddMediaChooser/RecropWarningModal/UnsavedChangesBlocker/PublishModal/
SignModal/CommentLibraryDrawer/SectionPickerModal/TagPickerModal); parent threads
state+handlers as explicit props. No logic, loader/action, or token changes.
