+++ app/components/editor/CommentLibraryDrawer.tsx
@@ -1,0 +1,188 @@
+import { formatRelativeTime } from "../../lib/format-relative-time";
+
+export interface CommentLibraryDrawerProps {
+ comments: {
+  filterMode: "auto" | "all";
+  setFilterMode: (m: "auto" | "all") => void;
+  sort: string;
+  setSort: (s: string) => void;
+  touchSnippet: (id: string) => void;
+ };
+ state: {
+  activeItem: { label?: unknown; name?: unknown } | null;
+  currentSection: { id: string; title?: string } | null;
+  activeItemId: string | null;
+  getResult: (itemId: string, sectionId?: string) => Record<string, unknown>;
+  getRatingLabel?: (ratingId: string | null | undefined) => string;
+  commentLibraryFilter: string;
+  setCommentLibraryFilter: (f: string) => void;
+  setCommentLibrarySelectedIdx: (i: number) => void;
+  commentLibrarySearch: string;
+  setCommentLibrarySearch: (s: string) => void;
+  commentLibrarySelectedIdx: number;
+  setShowCommentLibrary: (open: boolean) => void;
+ };
+ serverComments: Array<{ id: string; text: string; useCount?: number; lastUsedAt?: number | null }>;
+ onInsert: (sectionId: string, itemId: string, text: string) => void;
+ onClose: () => void;
+}
+
+export function CommentLibraryDrawer({ comments, state, serverComments, onInsert, onClose }: CommentLibraryDrawerProps) {
+ return (
+ <div className="fixed inset-0 z-[80] flex">
+ <div
+ className="absolute inset-0 bg-[rgba(15,23,42,0.4)] backdrop-blur-sm"
+ onClick={onClose}
+ />
+ <div className="relative ml-auto w-full max-w-md bg-ih-bg-card border-l border-ih-border shadow-ih-popover flex flex-col h-full">
+ <div className="flex items-center justify-between px-4 py-3 border-b border-ih-border">
+ <h3 className="text-[14px] font-bold">Comment Library</h3>
+ <button
+ onClick={onClose}
+ className="text-ih-fg-4 hover:text-ih-fg-2 text-lg"
+ >
+ &#x2715;
+ </button>
+ </div>
+
+ {/* Sort + Filter mode header */}
+ <div className="flex items-center gap-3 px-3 py-2 border-b border-ih-border">
+ <div className="flex items-center gap-1.5">
+ <span className="text-[10px] uppercase tracking-[0.1em] text-ih-fg-4">Filter</span>
+ <select
+ value={comments.filterMode}
+ onChange={e => comments.setFilterMode(e.target.value as 'auto' | 'all')}
+ className="px-2 py-1 rounded border border-ih-border bg-ih-bg-app text-[11px]"
+ >
+ <option value="auto">Auto</option>
+ <option value="all">All</option>
+ </select>
+ </div>
+ <div className="flex items-center gap-1.5 ml-auto">
+ <span className="text-[10px] uppercase tracking-[0.1em] text-ih-fg-4">Sort</span>
+ <select
+ value={comments.sort}
+ onChange={e => comments.setSort(e.target.value)}
+ className="px-2 py-1 rounded border border-ih-border bg-ih-bg-app text-[11px]"
+ >
+ <option value="relevance">Relevance</option>
+ <option value="recent">Recent use</option>
+ <option value="created">Recently added</option>
+ <option value="frequent">Most used</option>
+ <option value="alpha">A–Z</option>
+ </select>
+ </div>
+ </div>
+
+ {/* Context strip (auto mode + active item) */}
+ {comments.filterMode === 'auto' && state.activeItem && (
+ <div className="flex items-center gap-2 px-3 py-1.5 text-[11px] bg-ih-bg-muted border-b border-ih-border">
+ <span className="text-ih-fg-4">Context:</span>
+ <span>
+ {state.currentSection?.title} › {(state.activeItem.label || state.activeItem.name) as string}
+ </span>
+ {Boolean(state.activeItemId && state.getResult(state.activeItemId)?.rating) && (
+ <>
+ <span className="text-ih-fg-4">·</span>
+ <span>
+ {state.getRatingLabel?.(state.getResult(state.activeItemId as string)?.rating as string) ?? ''}
+ </span>
+ </>
+ )}
+ <button
+ onClick={() => comments.setFilterMode('all')}
+ className="ml-auto text-ih-fg-4 hover:text-ih-fg-2"
+ aria-label="Clear filter"
+ >×</button>
+ </div>
+ )}
+
+ {/* Filter chips */}
+ <div className="flex gap-1 px-4 py-2 border-b border-ih-border flex-wrap">
+ {[
+ { id: "all", label: "All" },
+ { id: "satisfactory", label: "Satisfactory" },
+ { id: "monitor", label: "Monitor" },
+ { id: "defect", label: "Defect" },
+ { id: "my-snippets", label: "My Snippets" },
+ ].map((f) => (
+ <button
+ key={f.id}
+ onClick={() => {
+ state.setCommentLibraryFilter(f.id);
+ state.setCommentLibrarySelectedIdx(0);
+ }}
+ className={`px-2.5 py-1 rounded-full text-[11px] font-bold ${
+ state.commentLibraryFilter === f.id
+ ? "bg-ih-primary-tint text-ih-primary"
+ : "text-ih-fg-4 hover:text-ih-fg-2"
+ }`}
+ >
+ {f.label}
+ </button>
+ ))}
+ </div>
+
+ {/* Search */}
+ <div className="px-4 py-2">
+ <input
+ id="comment-library-search"
+ type="text"
+ placeholder="Search comments..."
+ value={state.commentLibrarySearch}
+ onChange={(e) => {
+ state.setCommentLibrarySearch(e.target.value);
+ state.setCommentLibrarySelectedIdx(0);
+ }}
+ className="w-full px-3 py-2 rounded-md border border-ih-border bg-ih-bg-app text-[12px]"
+ autoFocus
+ />
+ <p className="text-[10px] text-ih-fg-4 mt-1">
+ {serverComments.length} comments
+ </p>
+ </div>
+
+ {/* Comment list (server-fetched, sort/filter aware) */}
+ <div className="flex-1 overflow-y-auto pb-2">
+ <ul className="divide-y divide-ih-border">
+ {serverComments.map((c, idx) => (
+ <li
+ key={c.id}
+ onClick={() => {
+ if (!state.currentSection || !state.activeItemId) return;
+ onInsert(
+ state.currentSection.id,
+ state.activeItemId,
+ c.text,
+ );
+ comments.touchSnippet(c.id);
+ state.setShowCommentLibrary(false);
+ }}
+ className={`cursor-pointer ${
+ idx === state.commentLibrarySelectedIdx
+ ? "bg-ih-primary-tint ring-1 ring-inset ring-ih-primary/30"
+ : ""
+ }`}
+ >
+ <div className="flex items-start gap-2 p-2.5 hover:bg-ih-bg-muted">
+ <p className="flex-1 text-[12px] text-ih-fg-2 leading-relaxed">
+ {c.text}
+ </p>
+ <span className="text-[10px] text-ih-fg-4 tabular-nums whitespace-nowrap">
+ {comments.sort === 'recent'   && c.lastUsedAt ? formatRelativeTime(c.lastUsedAt) : ''}
+ {comments.sort === 'frequent' && c.useCount   ? `${c.useCount}×`               : ''}
+ </span>
+ </div>
+ </li>
+ ))}
+ </ul>
+ {serverComments.length === 0 && (
+ <p className="text-[13px] text-ih-fg-3 text-center py-8">
+ No comments match the current filter.
+ </p>
+ )}
+ </div>
+ </div>
+ </div>
+ );
+}
